feat: 마이페이지 UI 개선 및 컴포넌트 분리 (#56)
- plane 관련 constants 파일 분리 및 설정\n- 마이페이지(Profile, MemberInfo, Plane) 영역별 컴포넌트 분리\n- 각 섹션별 UI 개선 및 코드 리팩토링\n- plane(종이비행기) 영역 UI 및 스크롤 개선
