+++ src/app/(header-layout)/mypage/components/memberinfo/MemberInfoSection.tsx
@@ -1,0 +1,43 @@
+import { MemberMeType } from '@/app/types/memberType'
+import { formatKoreanDate } from '@/utils/date'
+import MemberInfoItem from '@/app/(header-layout)/mypage/components/memberinfo/MemberInfoItem'
+
+interface MemberInfoSectionProps {
+  user: MemberMeType
+}
+
+export default function MemberInfoSection({ user }: MemberInfoSectionProps) {
+  return (
+    <section className="container px-4">
+      <div className="head5">
+        <p className="text-balck">나의 정보</p>
+      </div>
+      <div className="mt-3 w-full">
+        <ul className="space-y-4">
+          <MemberInfoItem
+            label="입사일"
+            value={user?.joinedAt && formatKoreanDate(user?.joinedAt)}
+          />
+          <MemberInfoItem label="생일" value={user?.birthday && formatKoreanDate(user?.birthday)} />
+          <MemberInfoItem
+            label="전화번호"
+            value={
+              <a className="text-purple-300" href={`tel:${user?.phone}`}>
+                {user?.phone}
+              </a>
+            }
+          />
+          <MemberInfoItem
+            label="이메일"
+            value={
+              <a className="text-purple-300" href={`mailto:${user?.email}`}>
+                {user?.email}
+              </a>
+            }
+          />
+          <MemberInfoItem label="깃허브" value={user?.githubId} />
+        </ul>
+      </div>
+    </section>
+  )
+}
