+++ src/app/(header-layout)/mypage/components/plane/PlaneMessageCard.tsx
@@ -1,0 +1,37 @@
+'use client'
+
+import Image from 'next/image'
+import IcoPlane from '@/assets/ico-plane.svg'
+import { PlaneMessage } from '@/app/(header-layout)/mypage/constants'
+import { formatRelativeTime } from '@/utils/date'
+interface PlaneMessageCardProps {
+  plane: PlaneMessage
+}
+
+export default function PlaneMessageCard({ plane }: PlaneMessageCardProps) {
+  return (
+    <div className="flex h-56 w-80 flex-col justify-between rounded-xl border bg-white p-5">
+      <p className="body2 line-clamp-4 text-gray-800">{plane?.message}</p>
+      <div className="flex items-end justify-between">
+        <div className="flex items-center gap-3">
+          <Image
+            src={plane.sender.profileImageUrl || '/default/avatar.png'}
+            alt={`${plane.sender.name} 프로필 이미지`}
+            width={52}
+            height={52}
+            className="min-h-13 rounded-full border border-gray-300"
+          />
+          <div className="flex flex-col gap-2 pt-2">
+            <p className="body2-sb text-black">{plane.sender.name}</p>
+            <p className="body4 text-gray-400">
+              {plane.sentAt && formatRelativeTime(plane.sentAt)}
+            </p>
+          </div>
+        </div>
+        <button type="button" className="flex-row-center h-10 w-10 rounded-full bg-purple-500">
+          <IcoPlane />
+        </button>
+      </div>
+    </div>
+  )
+}
