+++ src/app/(header-layout)/mypage/constants.ts
@@ -1,31 +1,45 @@
-export interface PlaneMessage {
+interface Sender {
   id: number
-  senderName: string
-  senderProfileImageUrl: string
+  name: string
+  profileImageUrl: string
+}
+
+export interface PlaneMessage {
+  messageId: number
+  sender: Sender
   message: string
   sentAt: string
 }
 
 export const planeMessages: PlaneMessage[] = [
   {
-    id: 1,
-    senderName: '김한솔',
-    senderProfileImageUrl: '/mock/avatar1.png',
-    message: '“정말 멋진 선임이에요! 정말 본받고 싶습니다..!!! 진심으로',
-    sentAt: '2025-06-13T14:20:00',
+    messageId: 1001,
+    sender: {
+      id: 1,
+      name: '김한솔',
+      profileImageUrl: 'https://cdn.kiring.app/profile/123.jpg',
+    },
+    message: '정말 멋진 선임이에요! 정말 본받고 싶습니다..!!! 진심으로',
+    sentAt: '2025-06-15T15:12:00',
   },
   {
-    id: 2,
-    senderName: '백혜인',
-    senderProfileImageUrl: '/mock/avatar2.png',
-    message: '덕분에 점심 잘 먹었어요!',
-    sentAt: '2025-06-12T11:05:00',
+    messageId: 1002,
+    sender: {
+      id: 2,
+      name: '백혜인',
+      profileImageUrl: 'https://cdn.kiring.app/profile/123.jpg',
+    },
+    message: '정말 킹 받네용 킹받',
+    sentAt: '2025-06-14T1:12:00',
   },
   {
-    id: 3,
-    senderName: '인범진',
-    senderProfileImageUrl: '/mock/avatar3.png',
-    message: '언제나 응원하고 있어요 ✨',
-    sentAt: '2025-06-10T08:40:00',
+    messageId: 1003,
+    sender: {
+      id: 3,
+      name: '인범진',
+      profileImageUrl: 'https://cdn.kiring.app/profile/123.jpg',
+    },
+    message: '응원합니다! :))',
+    sentAt: '2025-06-13T09:12:00',
   },
 ]
